requirements.txt
Adafruit_Blinka==5.2.3
adafruit_circuitpython_neopixel==6.0.0
board==1.0
